relocated the commands
