+++ commands/gary_channel_list.js
@@ -1,4 +1,4 @@
-const config = require('./config.json');
+const config = require('../config.json');
 const Discord = require('discord.js');
 const client = new Discord.Client();
