+++ commands/store_meme.js
